+++ lib/style/pipes.ex
@@ -114,7 +114,12 @@ defmodule Styler.Style.Pipes do
   end
 
   # a |> fun => a |> fun()
-  defp fix_pipe({:|>, meta, [lhs, {fun, fun_meta, nil}]}), do: {:|>, meta, [lhs, {fun, fun_meta, []}]}
+  defp fix_pipe({:|>, m, [lhs, {fun, m2, nil}]}), do: {:|>, m, [lhs, {fun, m2, []}]}
+  # rewrite anonymous function invocation to use `then/2`
+  # `a |> (& &1).() |> c` => `a |> then(& &1) |> c()`
+  # `a |> (fn x -> x end).() |> c` => `a |> then(fn x -> x end) |> c()`
+  defp fix_pipe({:|>, m, [lhs, {{:., m2, [{anon_fun, _, _}] = fun}, _, []}]}) when anon_fun in [:&, :fn],
+    do: {:|>, m, [lhs, {:then, m2, fun}]}
 
   # `lhs |> Enum.reverse() |> Enum.concat(enum)` => `lhs |> Enum.reverse(enum)`
   defp fix_pipe(